pyproject.toml
[project]
name = "multi-tool-agent"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "langchain-openai>=0.2",
    "langgraph>=0.2",
    "langgraph-checkpoint-redis>=0.1",
    "langchain-core>=0.3",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "simpleeval>=0.9",
    "duckduckgo-search>=6.3",
    "httpx>=0.27",
    "structlog>=24.4",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-exporter-otlp-proto-http>=1.27",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-semantic-conventions>=0.48b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx",
    "aiosqlite>=0.20",
]

[tool.uv]
dev-dependencies = ["pytest>=8.3", "pytest-asyncio>=0.24", "aiosqlite>=0.20"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
cache_dir = "/tmp/.pytest_cache"
